+ finish all requirements

+++ src/services/postgree/PlaylistsService.js
@@ -7,19 +7,13 @@ const AuthorizationError = require('../../exceptions/AuthorizationError');
 const {nanoid} = require('nanoid');
 
 class PlaylistsService {
-  constructor(collaborationService) {
+  constructor(collaborationService, cacheService) {
     this._pool = new Pool();
-
     this._collaborationService = collaborationService;
-
-    this.addPlaylist = this.addPlaylist.bind(this);
-    this.getPlaylists = this.getPlaylists.bind(this);
-    this.deletePlaylistById = this.deletePlaylistById.bind(this);
-    this.verifyPlaylistOwner = this.verifyPlaylistOwner.bind(this);
-    this.verifyPlaylistAccess = this.verifyPlaylistAccess.bind(this);
+    this._cacheService = cacheService;
   }
 
-  async addPlaylist({name, owner}) {
+  async addPlaylist(name, owner) {
     const id = `playlist-${nanoid(16)}`;
     const query = {
       text: 'INSERT INTO playlists VALUES($1, $2, $3) RETURNING id',
@@ -30,20 +24,33 @@ class PlaylistsService {
     if (!result.rows.length) {
       throw new InvariantError('Playlist gagal ditambahkan');
     }
+    await this._cacheService.delete(`playlists:${owner}`);
     return result.rows[0].id;
   }
 
   async getPlaylists(owner) {
-    const query = {
-      text: `SELECT playlists.id, playlists.name, users.username FROM playlists
+    try {
+      const result = await this._cacheService.get(`playlists:${owner}`);
+      return {
+        source: 'cache',
+        playlists: JSON.parse(result),
+      };
+    } catch (error) {
+      const query = {
+        text: `SELECT playlists.id, playlists.name, users.username FROM playlists
         LEFT JOIN users ON users.id = playlists.owner
         LEFT JOIN collaborations ON collaborations.playlist_id = playlists.id
         WHERE playlists.owner = $1 OR collaborations.user_id = $1`,
-      values: [owner],
-    };
+        values: [owner],
+      };
 
-    const result = await this._pool.query(query);
-    return result.rows;
+      const result = await this._pool.query(query);
+      await this._cacheService.set(`playlists:${owner}`, JSON.stringify(result.rows));
+      return {
+        source: 'database',
+        playlists: result.rows,
+      };
+    }
   }
 
   async getPlaylistById(id) {
@@ -63,25 +70,9 @@ class PlaylistsService {
     return result.rows[0];
   }
 
-  async editPlaylistById(id, {
-    title, year, performer, genre, duration,
-  }) {
-    const updatedAt = new Date().toISOString();
-    const query = {
-      text: 'UPDATE playlists SET title = $1, year = $2, performer = $3, genre = $4, duration = $5, updated_at = $6 WHERE id = $7 RETURNING id',
-      values: [title, year, performer, genre, duration, updatedAt, id],
-    };
-
-    const result = await this._pool.query(query);
-
-    if (!result.rows.length) {
-      throw new NotFoundError('Gagal memperbarui lagu. Id tidak ditemukan');
-    }
-  }
-
   async deletePlaylistById(id) {
     const query = {
-      text: 'DELETE FROM playlists WHERE id = $1 RETURNING id',
+      text: 'DELETE FROM playlists WHERE id = $1 RETURNING id, owner',
       values: [id],
     };
 
@@ -89,6 +80,9 @@ class PlaylistsService {
     if (!result.rows.length) {
       throw new InvariantError('Playlist gagal dihapus. Id tidak ditemukan');
     }
+
+    const {owner} = result.rows[0];
+    await this._cacheService.delete(`playlists:${owner}`);
   }
 
   async verifyPlaylistOwner(id, owner) {
